requirements.txt
praw==7.7.1
requests==2.31.0
python-dotenv==1.0.0
Pillow>=10.4.0
imagehash==4.3.1
opencv-python>=4.8.0
supabase==1.2.0